+++ src/VlasovSolver/schemes/upwind.jl
@@ -1,7 +1,8 @@
 @inline _upwind⁺(f, i, i⁻, c) = f[i] - c*(f[i] - f[i⁻])
 @inline _upwind⁻(f, i, i⁺, c) = f[i] - c*(f[i⁺] - f[i])
 
-function advect!(dest, src, ::Upwind, c, ws)
+function advect!(dest, src, scheme::Upwind, c, ws)
+    _validate(dest, src, scheme, ws)
     n = length(dest)
     if c > 0
         dest[1] = _upwind⁺(src, 1, n, c)